quiet installation of local extensions
Since many users ignore this notification, only 40% of users who noticed it managed to convert, and around 1% of all users never come back if they try to use we decided to quite this notitication on the connection. Instead it logs and we provide an explicit `Install Local Extensions` command to force sync at any time which will prompt a user.